Only write lookups for anchors w/ equivalent
Before this fix, the markfeaturewriter would just write lookups for all anchors found in all base glyphs.
If the data was clean, this was not a problem. However, in UFO subsets this could become an issue:
For example, a _belowLC lookup would be created, although there was no more combining mark to receive the belowLC anchor. This would result in an invalid mark feature file.

Fixes #14